Fixed potentially dangerous expression, vp[-fp->script->depth] Here, an unary minus was applied to unsigned integer value, with an unsigned result.  On machines where pointers are 64 bits and integers are 32-bits, this will not have the intended result.

+++ mozilla/js/src/jsinterp.c
@@ -708,7 +708,7 @@ out2:
                               cx->stackPool.current->base) <
                   JS_UPTRDIFF(cx->stackPool.current->avail,
                               cx->stackPool.current->base));
-        vp[-fp->script->depth] = (jsval)fp->pc;
+        vp[-(int)fp->script->depth] = (jsval)fp->pc;
     }
     return ok;
